+++ lib/mpi/messages/find_profile_message_helpers.rb
@@ -37,6 +37,54 @@ module MPI
         el << element('initialQuantity', value: 1)
       end
 
+      def build_gender
+        el = element('livingSubjectAdministrativeGender')
+        el << element('value', code: @gender)
+        el << element('semanticsText', text!: 'Gender')
+        el
+      end
+
+      def build_living_subject_birth_time
+        el = element('livingSubjectBirthTime')
+        el << element('value', value: Date.parse(@birth_date)&.strftime('%Y%m%d'))
+        el << element('semanticsText', text!: 'Date of Birth')
+        el
+      end
+
+      def build_living_subject_id
+        el = element('livingSubjectId')
+        el << element('value', root: '2.16.840.1.113883.4.1', extension: @ssn)
+        el << element('semanticsText', text!: 'SSN')
+        el
+      end
+
+      def build_living_subject_name
+        el = element('livingSubjectName')
+        value = element('value', use: 'L')
+        @given_names.each do |given_name|
+          value << element('given', text!: given_name)
+        end
+        value << element('family', text!: @family_name)
+        el << value
+        el << element('semanticsText', text!: 'Legal Name')
+        el
+      end
+
+      def build_patient_address
+        el = element('PatientAddress')
+        value = element('value', use: 'PHYS')
+        @street_address_lines.each do |street_address_line|
+          value << element('streetAddressLine', text!: street_address_line)
+        end
+        value << element('city', text!: @city)
+        value << element('state', text!: @state)
+        value << element('postalCode', text!: @postal_code)
+        value << element('country', text!: @country)
+        el << value
+        el << element('semanticsText', text!: 'Physical Address')
+        el
+      end
+
       def build_vba_orchestration
         el = element('otherIDsScopingOrganization')
         el << element('value', extension: 'VBA', root: '2.16.840.1.113883.4.349')
