+++ lib/mpi/messages/find_profile_message_with_address.rb
@@ -1,12 +1,24 @@
 # frozen_string_literal: true
 
+require_relative 'find_profile_message_fields'
+require_relative 'find_profile_message_with_address_fields'
+require_relative 'find_profile_message_helpers'
+require 'mpi/constants'
+
 module MPI
   module Messages
-    class FindProfileMessageWithAddress < FindProfileMessage
-      attr_reader :address
+    class FindProfileMessageWithAddress
+      include FindProfileMessageHelpers
+      attr_reader :address, :given_names, :family_name, :birth_date, :gender, :search_type
 
       def initialize(profile)
         required_fields_present?(profile)
+        @given_names = profile[:given_names]
+        @family_name = profile[:last_name]
+        @birth_date = profile[:birth_date]
+        # gender is optional and will default to nil if it DNE
+        @gender = profile[:gender]
+
         @address = profile[:address]
         @street_address_lines = [address['addressLine1'], address['addressLine2'], address['addressLine3']].compact
         @city = address['city']
@@ -15,7 +27,8 @@ module MPI
         @postal_code = address['zipCode5']
         @country = address['countryCodeISO2']
 
-        super(profile)
+        # Hardcode search_type for now, with the option to parameterize it later if needed
+        @search_type = MPI::Constants::CORRELATION_WITH_RELATIONSHIP_DATA
       end
 
       def required_fields_present?(profile)
@@ -48,24 +61,14 @@ module MPI
         el << build_gender if @gender.present?
         el << build_living_subject_birth_time
         el << build_living_subject_name
-        el << build_address
+        el << build_patient_address
         el << build_vba_orchestration if Settings.mvi.vba_orchestration
         el
       end
 
-      def build_address
-        el = element('PatientAddress')
-        value = element('value', use: 'PHYS')
-        @street_address_lines.each do |street_address_line|
-          value << element('streetAddressLine', text!: street_address_line)
-        end
-        value << element('city', text!: @city)
-        value << element('state', text!: @state)
-        value << element('postalCode', text!: @postal_code)
-        value << element('country', text!: @country)
-        el << value
-        el << element('semanticsText', text!: 'Physical Address')
-        el
+      def build_control_act_process
+        el = element('controlActProcess', classCode: 'CACT', moodCode: 'EVN')
+        el << element('code', code: 'PRPA_TE201305UV02', codeSystem: '2.16.840.1.113883.1.6')
       end
     end
   end
